+++ lib/twilio-ruby/rest/bulkexports/v1/export/export_custom_job.rb
@@ -11,8 +11,6 @@ module Twilio
     class Bulkexports < Domain
       class V1 < Version
         class ExportContext < InstanceContext
-          ##
-          # PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
           class ExportCustomJobList < ListResource
             ##
             # Initialize the ExportCustomJobList
@@ -147,8 +145,6 @@ module Twilio
             end
           end
 
-          ##
-          # PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
           class ExportCustomJobPage < Page
             ##
             # Initialize the ExportCustomJobPage
@@ -178,8 +174,6 @@ module Twilio
             end
           end
 
-          ##
-          # PLEASE NOTE that this class contains beta products that are subject to change. Use them with caution.
           class ExportCustomJobInstance < InstanceResource
             ##
             # Initialize the ExportCustomJobInstance