+++ lib/twilio-ruby/rest/verify/v2/service/entity/challenge.rb
@@ -46,8 +46,8 @@ module Twilio
               #   Challenge. Not shown to the end user. It must be a stringified JSON with only
               #   strings values eg. `{"ip": "172.168.1.234"}`
               # @param [String] auth_payload Optional payload used to verify the Challenge upon
-              #   creation. Only used with a Factor of type `totp` to carry an OTP used in the
-              #   verification.
+              #   creation. Only used with a Factor of type `totp` to carry the TOTP code that
+              #   needs to be verified.
               # @return [ChallengeInstance] Created ChallengeInstance
               def create(factor_sid: nil, expiration_date: :unset, details_message: :unset, details_fields: :unset, hidden_details: :unset, auth_payload: :unset)
                 data = Twilio::Values.of({