+++ spec/integration/api/v2010/account/call/feedback_spec.rb
@@ -9,6 +9,47 @@
 require 'spec_helper.rb'
 
 describe 'Feedback' do
+  it "can fetch" do
+    @holodeck.mock(Twilio::Response.new(500, ''))
+
+    expect {
+      @client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
+                       .calls('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
+                       .feedback().fetch()
+    }.to raise_exception(Twilio::REST::TwilioError)
+
+    expect(
+    @holodeck.has_request?(Holodeck::Request.new(
+        method: 'get',
+        url: 'https://api.twilio.com/2010-04-01/Accounts/ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Calls/CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Feedback.json',
+    ))).to eq(true)
+  end
+
+  it "receives fetch responses" do
+    @holodeck.mock(Twilio::Response.new(
+        200,
+      %q[
+      {
+          "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "date_created": "Thu, 20 Aug 2015 21:45:46 +0000",
+          "date_updated": "Thu, 20 Aug 2015 21:45:46 +0000",
+          "issues": [
+              "imperfect-audio",
+              "post-dial-delay"
+          ],
+          "quality_score": 1,
+          "sid": "CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+      }
+      ]
+    ))
+
+    actual = @client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
+                              .calls('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
+                              .feedback().fetch()
+
+    expect(actual).to_not eq(nil)
+  end
+
   it "can create" do
     @holodeck.mock(Twilio::Response.new(500, ''))
 
@@ -39,7 +80,7 @@ describe 'Feedback' do
               "imperfect-audio",
               "post-dial-delay"
           ],
-          "quality_score": 5,
+          "quality_score": 1,
           "sid": "CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
       }
       ]
@@ -52,62 +93,19 @@ describe 'Feedback' do
     expect(actual).to_not eq(nil)
   end
 
-  it "can fetch" do
-    @holodeck.mock(Twilio::Response.new(500, ''))
-
-    expect {
-      @client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
-                       .calls('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
-                       .feedback().fetch()
-    }.to raise_exception(Twilio::REST::TwilioError)
-
-    expect(
-    @holodeck.has_request?(Holodeck::Request.new(
-        method: 'get',
-        url: 'https://api.twilio.com/2010-04-01/Accounts/ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Calls/CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Feedback.json',
-    ))).to eq(true)
-  end
-
-  it "receives fetch responses" do
-    @holodeck.mock(Twilio::Response.new(
-        200,
-      %q[
-      {
-          "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-          "date_created": "Thu, 20 Aug 2015 21:45:46 +0000",
-          "date_updated": "Thu, 20 Aug 2015 21:45:46 +0000",
-          "issues": [
-              "imperfect-audio",
-              "post-dial-delay"
-          ],
-          "quality_score": 5,
-          "sid": "CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
-      }
-      ]
-    ))
-
-    actual = @client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
-                              .calls('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
-                              .feedback().fetch()
-
-    expect(actual).to_not eq(nil)
-  end
-
   it "can update" do
     @holodeck.mock(Twilio::Response.new(500, ''))
 
     expect {
       @client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
                        .calls('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
-                       .feedback().update(quality_score: 1)
+                       .feedback().update()
     }.to raise_exception(Twilio::REST::TwilioError)
 
-    values = {'QualityScore' => 1, }
     expect(
     @holodeck.has_request?(Holodeck::Request.new(
         method: 'post',
         url: 'https://api.twilio.com/2010-04-01/Accounts/ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Calls/CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX/Feedback.json',
-        data: values,
     ))).to eq(true)
   end
 
@@ -120,10 +118,9 @@ describe 'Feedback' do
           "date_created": "Thu, 20 Aug 2015 21:45:46 +0000",
           "date_updated": "Thu, 20 Aug 2015 21:45:46 +0000",
           "issues": [
-              "imperfect-audio",
-              "post-dial-delay"
+              "audio-latency"
           ],
-          "quality_score": 5,
+          "quality_score": 2,
           "sid": "CAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
       }
       ]
@@ -131,7 +128,7 @@ describe 'Feedback' do
 
     actual = @client.api.v2010.accounts('ACXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
                               .calls('CAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
-                              .feedback().update(quality_score: 1)
+                              .feedback().update()
 
     expect(actual).to_not eq(nil)
   end