+++ spec/integration/events/v1/event_type_spec.rb
@@ -58,6 +58,7 @@ describe 'EventType' do
                   "date_updated": "2020-08-13T13:28:20Z",
                   "type": "com.twilio.messaging.message.delivered",
                   "schema_id": "Messaging.MessageStatus",
+                  "public": true,
                   "description": "Messaging- delivered message",
                   "url": "https://events.twilio.com/v1/Types/com.twilio.messaging.message.delivered",
                   "links": {
@@ -69,6 +70,7 @@ describe 'EventType' do
                   "date_updated": "2020-08-13T13:28:19Z",
                   "type": "com.twilio.messaging.message.failed",
                   "schema_id": "Messaging.MessageStatus",
+                  "public": true,
                   "description": "Messaging- failed message",
                   "url": "https://events.twilio.com/v1/Types/com.twilio.messaging.message.failed",
                   "links": {
@@ -117,6 +119,7 @@ describe 'EventType' do
           "date_updated": "2020-08-13T13:28:20Z",
           "type": "com.twilio.messaging.message.delivered",
           "schema_id": "Messaging.MessageStatus",
+          "public": true,
           "description": "Messaging- delivered message",
           "url": "https://events.twilio.com/v1/Types/com.twilio.messaging.message.delivered",
           "links": {