+++ spec/integration/events/v1/schema/version_spec.rb
@@ -58,6 +58,7 @@ describe 'SchemaVersion' do
               {
                   "id": "Messaging.MessageStatus",
                   "schema_version": 1,
+                  "public": true,
                   "date_created": "2015-07-30T20:00:00Z",
                   "url": "https://events.twilio.com/v1/Schemas/Messaging.MessageStatus/Versions/1",
                   "raw": "https://events-schemas.twilio.com/Messaging.MessageStatus/1"
@@ -65,6 +66,7 @@ describe 'SchemaVersion' do
               {
                   "id": "Messaging.MessageStatus",
                   "schema_version": 2,
+                  "public": true,
                   "date_created": "2015-07-30T20:00:00Z",
                   "url": "https://events.twilio.com/v1/Schemas/Messaging.MessageStatus/Versions/2",
                   "raw": "https://events-schemas.twilio.com/Messaging.MessageStatus/2"
@@ -111,6 +113,7 @@ describe 'SchemaVersion' do
       {
           "id": "Messaging.MessageStatus",
           "schema_version": 1,
+          "public": true,
           "date_created": "2015-07-30T20:00:00Z",
           "url": "https://events.twilio.com/v1/Schemas/Messaging.MessageStatus/Versions/1",
           "raw": "https://events-schemas.twilio.com/Messaging.MessageStatus/1"