+++ spec/integration/verify/v2/service/entity/challenge_spec.rb
@@ -27,17 +27,17 @@ describe 'Challenge' do
     ))).to eq(true)
   end
 
-  it "receives create_without_auth_payload responses" do
+  it "receives create_push_without_auth_payload responses" do
     @holodeck.mock(Twilio::Response.new(
         201,
       %q[
       {
-          "sid": "YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "sid": "YC03aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "service_sid": "VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "entity_sid": "YEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "identity": "ff483d1ff591898a9942916050d2ca3f",
-          "factor_sid": "YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "factor_sid": "YF03aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z",
           "date_responded": "2015-07-30T20:00:00Z",
@@ -58,9 +58,9 @@ describe 'Challenge' do
               "ip": "172.168.1.234"
           },
           "factor_type": "push",
-          "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC03aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "links": {
-              "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
+              "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC03aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
           }
       }
       ]
@@ -73,17 +73,63 @@ describe 'Challenge' do
     expect(actual).to_not eq(nil)
   end
 
-  it "receives create_with_auth_payload responses" do
+  it "receives create_totp_without_auth_payload responses" do
     @holodeck.mock(Twilio::Response.new(
         201,
       %q[
       {
-          "sid": "YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "sid": "YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "service_sid": "VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "entity_sid": "YEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "identity": "ff483d1ff591898a9942916050d2ca3f",
-          "factor_sid": "YFaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "factor_sid": "YF02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "date_created": "2015-07-30T20:00:00Z",
+          "date_updated": "2015-07-30T20:00:00Z",
+          "date_responded": "2015-07-30T20:00:00Z",
+          "expiration_date": "2015-07-30T20:00:00Z",
+          "status": "pending",
+          "responded_reason": "none",
+          "details": {
+              "message": "Hi! Mr. John Doe, would you like to sign up?",
+              "date": "2020-07-01T12:13:14Z",
+              "fields": [
+                  {
+                      "label": "Action",
+                      "value": "Sign up in portal"
+                  }
+              ]
+          },
+          "hidden_details": {
+              "ip": "172.168.1.234"
+          },
+          "factor_type": "totp",
+          "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "links": {
+              "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
+          }
+      }
+      ]
+    ))
+
+    actual = @client.verify.v2.services('VAXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX') \
+                              .entities('identity') \
+                              .challenges.create(factor_sid: 'YFXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXXX')
+
+    expect(actual).to_not eq(nil)
+  end
+
+  it "receives create_totp_with_auth_payload responses" do
+    @holodeck.mock(Twilio::Response.new(
+        201,
+      %q[
+      {
+          "sid": "YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "service_sid": "VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "entity_sid": "YEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "identity": "ff483d1ff591898a9942916050d2ca3f",
+          "factor_sid": "YF02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "date_created": "2015-07-30T20:00:00Z",
           "date_updated": "2015-07-30T20:00:00Z",
           "date_responded": "2015-07-30T20:00:00Z",
@@ -103,10 +149,10 @@ describe 'Challenge' do
           "hidden_details": {
               "ip": "172.168.1.234"
           },
-          "factor_type": "push",
-          "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+          "factor_type": "totp",
+          "url": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
           "links": {
-              "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YCaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
+              "notifications": "https://verify.twilio.com/v2/Services/VAaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Entities/ff483d1ff591898a9942916050d2ca3f/Challenges/YC02aaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Notifications"
           }
       }
       ]